+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## CURRENT
 
+- Add Change.apply method for modifying objects
 - Update bunyan to 0.23.1
 - #143 Preserve raw Buffer value in Control objects
 - Test code coverage with node-istanbul