+++ CHANGES.md
@@ -29,6 +29,7 @@
   - Removed `spaced()` and `rndSpaced` from DN API
   - Fix parent/child rules regarding empty DNs
 - #254 Allow simple client bind with empty credentials
+- Add StartTLS support to client
 
 ## 0.7.1